refactor: вынес опции выбора в конфил для компонента RecipeFilterType
