+++ src/App/pages/RecipesPage/components/RecipeFilterType/RecipeFilterType.tsx
@@ -2,23 +2,7 @@ import cn from 'classnames';
 import { FC, useState } from 'react';
 import MultiDropdown, { MultiDropdownOption } from 'components/MultiDropdown';
 import { useRecipesContext } from 'context/RecipesContext';
-
-const optionType: MultiDropdownOption[] = [
-  { key: 'main course', value: 'main course' },
-  { key: 'side dish', value: 'side dish' },
-  { key: 'dessert', value: 'dessert' },
-  { key: 'appetizer', value: 'appetizer' },
-  { key: 'salad', value: 'salad' },
-  { key: 'bread', value: 'bread' },
-  { key: 'breakfast', value: 'breakfast' },
-  { key: 'soup', value: 'soup' },
-  { key: 'beverage', value: 'beverage' },
-  { key: 'sauce', value: 'sauce' },
-  { key: 'marinade', value: 'marinade' },
-  { key: 'fingerfood', value: 'fingerfood' },
-  { key: 'snack', value: 'snack' },
-  { key: 'drink', value: 'drink' },
-];
+import { optionType } from './config';
 
 type RecipeFilterTypePorps = {
   className?: string;
