+++ src/App/pages/RecipesPage/components/RecipeFilterType/config.ts
@@ -1,0 +1,20 @@
+import { MultiDropdownOption } from 'components/MultiDropdown';
+
+const optionType: MultiDropdownOption[] = [
+  { key: 'main course', value: 'main course' },
+  { key: 'side dish', value: 'side dish' },
+  { key: 'dessert', value: 'dessert' },
+  { key: 'appetizer', value: 'appetizer' },
+  { key: 'salad', value: 'salad' },
+  { key: 'bread', value: 'bread' },
+  { key: 'breakfast', value: 'breakfast' },
+  { key: 'soup', value: 'soup' },
+  { key: 'beverage', value: 'beverage' },
+  { key: 'sauce', value: 'sauce' },
+  { key: 'marinade', value: 'marinade' },
+  { key: 'fingerfood', value: 'fingerfood' },
+  { key: 'snack', value: 'snack' },
+  { key: 'drink', value: 'drink' },
+];
+
+export { optionType };
